WIP: Windows support.
Strictly speaking this is support for building with clang against the
mingw libraries/headers to cross compile for Windows, but it does
affect the native Windows compilation as well.

This is mostly support for win32 threads, but includes fixes for other
parts of POSIX that we use that aren't provided by Windows (such as
nanosleep).

When using mingw, the C++ ABI comes from libsupc++. Figuring out the
native Windows libraries is the next step.

Exceptions are not working yet.